:sparkles: Implementando a retirada do processo de pauta

+++ src/app/modules/acervo/acoes/acoes.component.ts
@@ -2,7 +2,8 @@ import { HttpClient } from '@angular/common/http';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { FuseAlertService } from '@fuse/components/alert';
-import { Processo } from '../tabela/tabela.component';
+import { ProcessoService } from 'app/modules/services/processo.service';
+import { Processo, SituacaoDoProcesso } from '../tabela/tabela.component';
 import { AgruparEmlistaComponent } from './agrupar-emlista/agrupar-emlista.component';
 import { AlertaComponent } from './agrupar-emlista/gerenciar-listas/alerta/alerta.component';
 import { PautarComponent } from './pautar/pautar.component';
@@ -14,6 +15,11 @@ import { PautarComponent } from './pautar/pautar.component';
 })
 export class AcoesComponent implements OnInit {
 
+  public alerta: {
+    titulo: string,
+    mensagem: string,
+  }
+
   mobile: boolean;
   @Output() Allselected = new EventEmitter();
   @Output() colecaoIdsDasTags = new EventEmitter<Array<{id: number}>>();
@@ -24,6 +30,7 @@ export class AcoesComponent implements OnInit {
     private _httpClient: HttpClient,
     private _matDialog: MatDialog, 
     private _fuseAlertService: FuseAlertService,
+    private _processoService: ProcessoService,
   ) {
     if (document.body.clientWidth <= 800) {
       this.mobile = true
@@ -35,6 +42,10 @@ export class AcoesComponent implements OnInit {
 
 
   ngOnInit(): void {
+    this.alerta = {
+      titulo: 'Erro de validação',
+      mensagem: 'Selecione um ou mais processos',
+    }
   }
 
   selectAll(completed) {
@@ -100,40 +111,63 @@ export class AcoesComponent implements OnInit {
   }
 
   retirarDePauta(): void {
-    console.log(this.processos);
-    const dialogRef = this._matDialog.open(AlertaComponent, {
-      data: {
-        titulo: 'Retirar processo',
-        mensagem: `Você tem certeza que deseja retirar esse(s) processo(s) de pauta? 
-                  ${this.exibeDescricaoDosProcessos()}
-                  `,
-      }
-    });
+    if (!this.verificaProcesso()) {
+      this.mostrarAlerta();
+    } else if (this.processos.some(p => p.situacao !== SituacaoDoProcesso.Pautado)) {
+      this.alertaDeErro('Erro nos processos selecionados', 'Selecione apenas os processos com a situação: Pautado.');
+    } else {
+      const dialogRef = this._matDialog.open(AlertaComponent, {
+        data: {
+          titulo: 'Retirar processo',
+          mensagem: `Você tem certeza que deseja retirar esse(s) processo(s) de pauta? 
+                    ${this.exibeDescricaoDosProcessos()}
+                    `,
+        }
+      });
 
-    dialogRef.afterClosed().subscribe(resultado => {
-      if (resultado === 'ok') {
-        // DELETE
-        this.processos.forEach(({id}) => {
-          this._httpClient.delete(`processos/${id}/pautar`).subscribe({
-            next: () => {
-              console.log('excluido');
-            }
+      dialogRef.afterClosed().subscribe(resultado => {
+        if (resultado === 'ok') {
+          // DELETE
+          this.processos.forEach(({id}) => {
+            this._httpClient.delete(`processos/${id}/pautar`).subscribe({
+              next: () => {
+                this._processoService.setProcessosRemovidosDaPauta(true);
+
+                this._fuseAlertService.show('sucesso');
+                setTimeout(() => {
+                  this._fuseAlertService.dismiss('sucesso');
+                }, 5000);
+              }
+            });
           });
-        });
-      } else {
-        dialogRef.close();
-      }
-    });
+        } else {
+          dialogRef.close();
+        }
+      });
+    }
   }
 
   exibeDescricaoDosProcessos(): string {
-    let descricoes = ``;
+    let descricoes = `\n`;
 
     this.processos.forEach(processo => {
-      descricoes += `\n${processo.classe} ${processo.numero} - ${processo.descricao}\n`;
+      descricoes += `${processo.classe} ${processo.numero} ${processo.nome}\n`;
     });
 
     return descricoes;
+  }
+
+  alertaDeErro(titulo: string, mensagem: string): void {
+    this.alerta = {
+      titulo,
+      mensagem,
+    };
+
+    this._fuseAlertService.show('alertBox');
+
+    setTimeout(() => {
+      this._fuseAlertService.dismiss('alertBox');
+    }, 5000);
   }
 
 }
